+++ graphistry/compute/predicates/ASTPredicate.py
@@ -1,0 +1,6 @@
+class ASTPredicate():
+    """
+    Internal, not intended for use outside of this module.
+    These are fancy columnar predicates used in {k: v, ...} node/edge df matching when going beyond primitive equality
+    """
+    pass
